// src/chat/interfaces/request-body.interface.ts
// Generated by https://quicktype.io

export interface RequestBody {
    model:       string;
    prompt:      string;
    temperature: number;
    max_tokens:  number;
    stop:        string[];
    messages:    Message[];
}

export interface Message {
    role:    Role;
    content: string;
}

export enum Role {
    Assistant = "assistant",
    System = "system",
    User = "user",
}